backend/requirements.txt
python-telegram-bot>=22.0
claude-agent-sdk>=0.1.50
aiomysql>=0.2.0
aiohttp>=3.9.0
structlog>=24.0.0
pydantic-settings>=2.0.0
playwright>=1.40.0
fastapi>=0.115.0
uvicorn>=0.32.0
sse-starlette>=2.0.0
telegramify-markdown>=0.1.3
openai>=1.40.0
tiktoken>=0.7.0
anthropic>=0.40.0
pdfplumber>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.0
aiofiles>=24.1.0
httpx>=0.27
